// Server/src/notifications/dto/notification.enum.ts
export enum NotificationType {
  FR_REQ = 'FRIEND_REQUEST',
  FR_ACC = 'FRIEND_ACCEPTED',
  FR_REF = 'FRIEND_REFUSED',
  CH_ADDUSER_REQ = 'CHANNEL_ADDUSER_REQUEST',
  CH_ADDUSER_ACC = 'CHANNEL_ADDUSER_ACCEPTED',
  CH_ADDUSER_REF = 'CHANNEL_ADDUSER_REFUSED',
  POS_REAC = 'POSITIVE_REACTION',
  NEG_REAC = 'NEGATIVE_REACTION',
  NEW_COMM = 'NEW_COMMENT',
  SMM_REQ = 'SMM_REQUEST',
  SMM_ACC = 'SMM_ACCEPTED',
  SMM_REF = 'SMM_REFUSED',
  REV = 'REVIEW',
  MEN = 'MENTION'
}